--- my-app/src/utils/constants/data.tsx
@@ -1,8 +1,0 @@
-
-import ImagePath from ".";
-export const iconsData = [
-  { src: ImagePath.image, alt: 'Image not found' },
-  { src: ImagePath.image1, alt: 'Image not found' },
-  { src: ImagePath.image2, alt: 'Image not found' },
-  { src: ImagePath.image3, alt: 'Image not found' }
-];
